requirements.txt
faster-whisper==1.2.0
librosa==0.11.0
OpenCC==1.1.9
opuslib==3.0.1
scikit-learn==1.7.2
torch==2.9.0
torchaudio==2.9.0
websocket-client==1.9.0
websocket-server==0.6.4
pydub==0.25.1
pyaudio==0.2.14
modelscope==1.31.0
addict==2.4.0
datasets==3.6.0
pillow==12.0.0
simplejson==3.20.2
sortedcontainers==2.4.0
